refactor: reorganize views
