+++ lib/main.dart
@@ -2,13 +2,13 @@ import 'package:fauth/api/ccid_fido_api.dart';
 import 'package:fauth/repositories/credential_repository.dart';
 import 'package:fauth/viewmodels/keys_viewmodel.dart';
 import 'package:fauth/viewmodels/navigation_viewmodel.dart';
-import 'package:fauth/views/logger/models/log_viewmodel.dart';
+import 'package:fauth/viewmodels/log_viewmodel.dart';
 import 'package:fauth/common/app_logger.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:window_manager/window_manager.dart';
 import 'package:fauth/common/system.dart' as system;
-import 'package:fauth/views/home/home_page.dart';
+import 'package:fauth/views/home.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
