+++ lib/viewmodels/log_viewmodel.dart
