+++ lib/widgets/adaptive_scaffold.dart
